requirements.txt
chromadb==1.3.5
faiss-cpu==1.13.0
matplotlib==3.10.7
numpy==2.3.5
openai==2.8.1
pandas==2.3.3
pinecone==8.0.0
pymilvus==2.6.3
python-dotenv==1.2.1
scikit-learn==1.7.2
seaborn==0.13.2
sentence-transformers==5.1.2
torch==2.7.1+cu118
tqdm==4.67.1
psutil==7.1.3
